pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autovid"
version = "0.1.0"
description = "Agentic, model-pluggable short-form video editing platform"
readme = "SPEC.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # ─── Agent Framework ────────────────────────────────
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",

    # ─── API ────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "python-multipart>=0.0.12",

    # ─── Task Queue ─────────────────────────────────────
    "arq>=0.26.0",
    "redis>=5.0.0",

    # ─── Database ───────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # ─── Object Storage ────────────────────────────────
    "minio>=7.2.0",

    # ─── ASR ────────────────────────────────────────────
    "faster-whisper>=1.1.0",

    # ─── Data / Schemas ─────────────────────────────────
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",

    # ─── HTTP Client ────────────────────────────────────
    "httpx>=0.27.0",

    # ─── Audio Analysis ─────────────────────────────────
    "librosa>=0.10.2",
    "pyloudnorm>=0.1.1",

    # ─── Subtitle ───────────────────────────────────────
    "pysrt>=1.1.2",

    # ─── Container Management ───────────────────────────
    "docker>=7.1.0",

    # ─── Observability ──────────────────────────────────
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",

    # ─── Social APIs ────────────────────────────────────
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

vllm = [
    "vllm>=0.6.0",
]

nat = [
    "nvidia-nat>=1.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/autovid"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
